requirements.txt
python-dotenv
discord.py
notion-client
python-dateutil
pytz
openai
filelock